+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added Funky Locomotion integration (raoulvdberge)
 - Minor fixes to autocrafting (raoulvdberge)
 - Fixed Exporter in Regulator Mode not regulating properly when same item is specified multiple times (raoulvdberge)
+- Fixed air appearing in Grid (raoulvdberge)
 
 ### 1.5.25
 - Fixed not being able to autocraft different Storage Drawers' wood drawers (raoulvdberge)